refactoring/slice
