+++ vite-project/src/redux/invoices/all/selectors.js
@@ -1,0 +1,3 @@
+export const selectAllInvoices = (state) => state.allInvoices.allInvoices;
+export const selectIsLoading = (state) => state.allInvoices.isLoading;
+export const selectError = (state) => state.allInvoices.error
