--- vite-project/src/redux/invoices/selectors.js
@@ -1,1 +1,0 @@
-export const selectAllInvoices = (state) => state.allInvoices.allInvoices;
