+++ vite-project/src/redux/invoices/single/selectors.js
@@ -1,0 +1,4 @@
+export const selectInvoice = (state) => state.invoice.invoice;
+export const selectEditInvoice = (state) => state.invoice.editInvoice;
+export const selectIsLoading = (state) => state.invoice.isLoading;
+export const selectError = (state) => state.invoice.error
